Fix namespace, move var_dump before exception code
